+++ CHANGES.md
@@ -1,3 +1,8 @@
+#### 0.0.9 / 2022-07-29
+* Plugin rename to "Super Simple Cookie Bar"
+* text-domain changed to "super-simple-cookie-bar"
+* Hosted CSS and JS locally
+
 #### 0.0.8 / 2022-07-29 =
 * text-domain changed to "super-simple-cookie-bar-by-ncompass"